Detalles esteticos
Se creo una pantalla extra al momento de eliminar un usuario para que el usuario confirme que deasea eliminarlo. Es decir, ya no se eliminara en una sola instancia si no que el usuario debera confirmar que desea hacerlo.
También se modifico minimamente algunos detalles esteticos en otras vistas.

+++ SistemaEncomiendas/aplicaciones/moduloUsuarios/views.py
@@ -64,5 +64,7 @@ def verUsuario(request, pk):
 #eliminar
 def eliminar_usuario(request, pk):
     user = get_object_or_404(CustomUser, idUsuario=pk)
-    user.delete()
-    return redirect('/usuarios/')
+    if request.method == 'POST':
+        user.delete()
+        return redirect('/usuarios/')
+    return render(request, 'moduloUsuarios/eliminarUsuario.html',{'user':user})
